pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ghaiw"
version = "0.1.0"
description = "AI-agent-driven git workflow management CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Ivan Viragine" },
]
keywords = ["git", "workflow", "ai", "cli", "github"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlmodel>=0.0.16",
    "pyyaml>=6.0",
    "rich>=13.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
    "pre-commit>=3.7",
]

[project.scripts]
ghaiw = "ghaiw.cli.main:app"

[project.urls]
Homepage = "https://github.com/ivanviragine/ghaiw-py"
Repository = "https://github.com/ivanviragine/ghaiw-py"

[tool.hatch.build.targets.wheel]
packages = ["src/ghaiw"]

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["ghaiw"]
omit = ["tests/*"]
